feat: implement pagination for x/wasm query (#85)
* feat: add legacy_querier

* feat: add pagination query for contract history

* feat: add pagination for rest api

* fix: modify unit test

* fix: apply review

* feat: add pagination for queryContractListByCode, queryCodeList

* fix: remove legacy_querier

* fix: fix integration test error

* fix: apply review

* refactor: move filteredPaginate to keeper/querier.go

* fix: fix offset condition

* fix: fix lint error

+++ x/wasm/internal/keeper/test_fuzz.go
@@ -17,7 +17,7 @@ func FuzzAddr(m *sdk.AccAddress, c fuzz.Continue) {
 }
 
 func FuzzAbsoluteTxPosition(m *types.AbsoluteTxPosition, c fuzz.Continue) {
-	m.BlockHeight = int64(c.RandUint64()) // can't be negative
+	m.BlockHeight = c.RandUint64() // can't be negative
 	m.TxIndex = c.RandUint64()
 }
